Componente especifico de render de la imagen
